requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.7.9
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
distro==1.9.0
Django==5.2.4
django-browser-reload==1.18.0
django-markdownify==0.9.5
django-storages==1.14.6
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.97.1
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
tinycss2==1.4.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
